requirements.txt
pandas
requests
cloudscraper
flask
google-genai